move filter related code out of Models and into Filtering + tidy ups
