+++ src/FamilyHubs.ServiceDirectory.Web/Filtering/FilterAspect.cs
@@ -1,7 +1,7 @@
-﻿using FamilyHubs.ServiceDirectory.Web.Models.Interfaces;
+﻿using FamilyHubs.ServiceDirectory.Web.Filtering.Interfaces;
 using System.Diagnostics;
 
-namespace FamilyHubs.ServiceDirectory.Web.Models;
+namespace FamilyHubs.ServiceDirectory.Web.Filtering;
 
 [DebuggerDisplay("{Id}")]
 public sealed record FilterAspect(string Id, string Description, bool SelectedByDefault = false) : IFilterAspect;
