+++ src/FamilyHubs.ServiceDirectory.Web/Filtering/Interfaces/FilterType.cs
@@ -1,0 +1,7 @@
+﻿namespace FamilyHubs.ServiceDirectory.Web.Filtering.Interfaces;
+
+public enum FilterType
+{
+    Checkboxes,
+    Radios
+}
