+++ src/FamilyHubs.ServiceDirectory.Web/Pages/ServiceFilter/Index.cshtml
@@ -50,7 +50,7 @@ else {
                         </div>
 
                         <div class="moj-filter__heading-action">
-                            <p><button type="submit" class="govuk-link app-button-link" name="remove" value="@PostFilter.RemoveAll">Clear filters</button></p>
+                            <p><button type="submit" class="govuk-link app-button-link" name="remove" value="@IFilter.RemoveAll">Clear filters</button></p>
                         </div>
                     </div>
